Added file with tests. Fixed classes (added verification)

+++ task02сonsole/src/by/training/03random/Randomizer.java
@@ -4,11 +4,11 @@ public class Randomizer{
     public static void main(String[] args){
         Scanner in = new Scanner(System.in);
         System.out.print("Enter count of numbers: ");
-        int count = in.nextInt();
+        short count = in.nextShort();
         System.out.println("1- with a new line; 2- without a new line");
         int presenceOfNewLine = in.nextInt();
 
-        for (int i=0; i<count; i++){
+        for (short i=0; i<count; i++){
             int number = (int) (Math.random()*100);
             if (presenceOfNewLine==1){
                 System.out.println(number);
