+++ task02сonsole/src/by/training/06time/TimeAssignment.java
@@ -9,7 +9,7 @@ public class TimeAssignment {
         System.out.print("Enter your last name: ");
         String lastName = in.nextLine();
         System.out.print("Enter count of days you need to do this task: ");
-        int neededDays = in.nextInt();
+        short neededDays = in.nextShort();
 
         Calendar calendar = new GregorianCalendar();
         Date currentTime = calendar.getTime();
@@ -20,3 +20,4 @@ public class TimeAssignment {
         System.out.println("Deadline: "+calendar.getTime());
     }
 }
+
